# Swagger/API-Documentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Game Data API",
    "description": "List data Game Suite's API",
    "contact": {
      "email": "divajuni0406@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "localhost"
    }
  ],
...
  "paths": {
...
    "/master-data-admin": {
      "get": {
        "tags": ["Admin Only!"],
        "summary": "Master Data Admin For Admin Page Rendering",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful to render"
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
    "/master-data-user": {
      "get": {
        "tags": ["Admin Only!"],
        "summary": "Master Data Admin For User Page Rendering",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful to render"
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT Authorization For CRUD Data",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}